// frontend/src/client/models/Recipe.ts
// tslint:disable
/**
 * Macro Macro API
 * An API that provides access to recipe and ingredient information, both user-generated and sourced from the USDA
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    AmountOfNamedMacros,
    AmountOfNamedMacrosFromJSON,
    AmountOfNamedMacrosFromJSONTyped,
    AmountOfNamedMacrosToJSON,
    NamedMacros,
    NamedMacrosFromJSON,
    NamedMacrosFromJSONTyped,
    NamedMacrosToJSON,
} from './';

/**
 * 
 * @export
 * @interface Recipe
 */
export interface Recipe extends NamedMacros {
    /**
     * 
     * @type {Array<AmountOfNamedMacros>}
     * @memberof Recipe
     */
    foods: Array<AmountOfNamedMacros>;
    /**
     * 
     * @type {number}
     * @memberof Recipe
     */
    portionSize: number;
}

export function RecipeFromJSON(json: any): Recipe {
    return RecipeFromJSONTyped(json, false);
}

export function RecipeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Recipe {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        ...NamedMacrosFromJSONTyped(json, ignoreDiscriminator),
        'foods': (json['foods'] as Array<any>).map(AmountOfNamedMacrosFromJSON),
        'portionSize': json['portionSize'],
    };
}

export function RecipeToJSON(value?: Recipe): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        ...NamedMacrosToJSON(value),
        'foods': (value.foods as Array<any>).map(AmountOfNamedMacrosToJSON),
        'portionSize': value.portionSize,
    };
}